refactor: extract prefs and sorting logic
