+++ lib/screens/v2/training_pack_template_list_core.dart
@@ -1,27 +1,5 @@
 part of 'training_pack_template_list_screen.dart';
 
-class TrainingPackTemplatePrefs {
-  static const hideCompleted = 'tpl_hide_completed';
-  static const groupByStreet = 'tpl_group_by_street';
-  static const groupByType = 'tpl_group_by_type';
-  static const mixedHandGoalOnly = 'tpl_mixed_handgoal_only';
-  static const mixedCount = 'tpl_mixed_count';
-  static const mixedStreet = 'tpl_mixed_street';
-  static const mixedAuto = 'tpl_mixed_auto';
-  static const endlessDrill = 'tpl_endless_drill';
-  static const favorites = 'tpl_favorites';
-  static const showFavoritesOnly = 'tpl_show_fav_only';
-  static const sortOption = 'tpl_sort_option';
-  static const stackFilter = 'tpl_stack_filter';
-  static const posFilter = 'tpl_pos_filter';
-  static const difficultyFilter = 'tpl_diff_filter';
-  static const streetFilter = 'tpl_street_filter';
-  static const recentPacks = 'tpl_recent_packs';
-  static const inProgress = 'tpl_in_progress';
-  static const mixedLastRun = 'tpl_mixed_last_run';
-  static const continueLast = 'tpl_continue_last';
-}
-
 class TrainingPackTemplateListScreen extends StatefulWidget {
   const TrainingPackTemplateListScreen({super.key});
 
@@ -35,9 +13,8 @@ class TrainingPackTemplateListScreen extends StatefulWidget {
 
 class _TrainingPackTemplateListScreenState
     extends State<TrainingPackTemplateListScreen>
-    with TrainingPackTemplateIo, TrainingPackTemplateFilterPanel {
+    with TrainingPackTemplateIo, TrainingPackTemplateFilterPanel, TrainingPackTemplateSortFilter {
   static const _stackRanges = ['0-9', '10-15', '16-25', '26+'];
-  final List<TrainingPackTemplate> _templates = [];
   bool _loading = false;
   String _query = '';
   late TextEditingController _searchCtrl;
@@ -52,7 +29,6 @@ class _TrainingPackTemplateListScreenState
   bool _groupByStreet = false;
   bool _groupByType = false;
   bool _icmOnly = false;
-  String _sort = 'coverage';
   List<GeneratedPackInfo> _history = [];
   int _mixedCount = 20;
   bool _mixedAutoOnly = false;
@@ -73,31 +49,7 @@ class _TrainingPackTemplateListScreenState
   bool _autoEvalRunning = false;
   Timer? _autoEvalTimer;
   bool _autoEvalQueued = false;
-  String? _stackFilter;
-  HeroPosition? _posFilter;
-  String? _difficultyFilter;
-  String? _streetFilter;
   final List<String> _recentIds = [];
-
-  Future<void> _loadSort() async {
-    final prefs = await SharedPreferences.getInstance();
-    final value = prefs.getString(TrainingPackTemplatePrefs.sortOption);
-    if (mounted && value != null) {
-      setState(() {
-        _sort = value;
-        _sortTemplates();
-      });
-    }
-  }
-
-  Future<void> _setSort(String value) async {
-    setState(() {
-      _sort = value;
-      _sortTemplates();
-    });
-    final prefs = await SharedPreferences.getInstance();
-    await prefs.setString(TrainingPackTemplatePrefs.sortOption, value);
-  }
 
   Future<void> refreshFromStorage() async {
     final list = await TrainingPackStorage.load();
@@ -121,59 +73,6 @@ class _TrainingPackTemplateListScreenState
     final list = map.values.toList()
       ..sort((a, b) => b.ts.compareTo(a.ts));
     return list;
-  }
-
-  void _sortTemplates() {
-    switch (_sort) {
-      case 'coverage':
-        _templates.sort((a, b) {
-          double cov(TrainingPackTemplate t) => t.coveragePercent ?? -1;
-          final r = cov(b).compareTo(cov(a));
-          return r == 0
-              ? a.name.toLowerCase().compareTo(b.name.toLowerCase())
-              : r;
-        });
-        break;
-      case 'created':
-        _templates.sort((a, b) {
-          final r = b.createdAt.compareTo(a.createdAt);
-          return r == 0
-              ? a.name.toLowerCase().compareTo(b.name.toLowerCase())
-              : r;
-        });
-        break;
-      case 'spots':
-        _templates.sort((a, b) {
-          final r = b.spots.length.compareTo(a.spots.length);
-          return r == 0
-              ? a.name.toLowerCase().compareTo(b.name.toLowerCase())
-              : r;
-        });
-        break;
-      case 'tag':
-        _templates.sort((a, b) {
-          final tagA = a.tags.isNotEmpty ? a.tags.first.toLowerCase() : '';
-          final tagB = b.tags.isNotEmpty ? b.tags.first.toLowerCase() : '';
-          final r = tagA.compareTo(tagB);
-          return r == 0
-              ? a.name.toLowerCase().compareTo(b.name.toLowerCase())
-              : r;
-        });
-        break;
-      case 'last_trained':
-        _templates.sort((a, b) {
-          final aDt = a.lastTrainedAt ?? DateTime.fromMillisecondsSinceEpoch(0);
-          final bDt = b.lastTrainedAt ?? DateTime.fromMillisecondsSinceEpoch(0);
-          final r = bDt.compareTo(aDt);
-          return r == 0
-              ? a.name.toLowerCase().compareTo(b.name.toLowerCase())
-              : r;
-        });
-        break;
-      default:
-        _templates.sort(
-            (a, b) => a.name.toLowerCase().compareTo(b.name.toLowerCase()));
-    }
   }
 
   Future<void> _loadProgress() async {
@@ -426,11 +325,7 @@ class _TrainingPackTemplateListScreenState
     await prefs.setBool(TrainingPackTemplatePrefs.showFavoritesOnly, value);
   }
 
-  Future<void> _setShowInProgressOnly(bool value) async {
-    setState(() => _showInProgressOnly = value);
-    final prefs = await SharedPreferences.getInstance();
-    await prefs.setBool(TrainingPackTemplatePrefs.inProgress, value);
-  }
+  
 
   Future<void> _saveFavorites() async {
     final prefs = await SharedPreferences.getInstance();
@@ -445,80 +340,6 @@ class _TrainingPackTemplateListScreenState
       }
     });
     await _saveFavorites();
-  }
-
-  Future<void> _loadStackFilter() async {
-    final prefs = await SharedPreferences.getInstance();
-    if (mounted)
-      setState(() =>
-          _stackFilter = prefs.getString(TrainingPackTemplatePrefs.stackFilter));
-  }
-
-  Future<void> _setStackFilter(String? value) async {
-    setState(() => _stackFilter = value);
-    final prefs = await SharedPreferences.getInstance();
-    if (value == null) {
-      await prefs.remove(TrainingPackTemplatePrefs.stackFilter);
-    } else {
-      await prefs.setString(TrainingPackTemplatePrefs.stackFilter, value);
-    }
-  }
-
-  Future<void> _loadPosFilter() async {
-    final prefs = await SharedPreferences.getInstance();
-    final name = prefs.getString(TrainingPackTemplatePrefs.posFilter);
-    if (mounted) {
-      setState(() => _posFilter = name == null
-          ? null
-          : HeroPosition.values.firstWhere(
-              (e) => e.name == name,
-              orElse: () => HeroPosition.sb,
-            ));
-    }
-  }
-
-  Future<void> _loadDifficultyFilter() async {
-    final prefs = await SharedPreferences.getInstance();
-    if (mounted)
-      setState(() => _difficultyFilter =
-          prefs.getString(TrainingPackTemplatePrefs.difficultyFilter));
-  }
-
-  Future<void> _setPosFilter(HeroPosition? value) async {
-    setState(() => _posFilter = value);
-    final prefs = await SharedPreferences.getInstance();
-    if (value == null) {
-      await prefs.remove(TrainingPackTemplatePrefs.posFilter);
-    } else {
-      await prefs.setString(TrainingPackTemplatePrefs.posFilter, value.name);
-    }
-  }
-
-  Future<void> _setDifficultyFilter(String? value) async {
-    setState(() => _difficultyFilter = value);
-    final prefs = await SharedPreferences.getInstance();
-    if (value == null) {
-      await prefs.remove(TrainingPackTemplatePrefs.difficultyFilter);
-    } else {
-      await prefs.setString(TrainingPackTemplatePrefs.difficultyFilter, value);
-    }
-  }
-
-  Future<void> _loadStreetFilter() async {
-    final prefs = await SharedPreferences.getInstance();
-    if (mounted)
-      setState(() =>
-          _streetFilter = prefs.getString(TrainingPackTemplatePrefs.streetFilter));
-  }
-
-  Future<void> _setStreetFilter(String? value) async {
-    setState(() => _streetFilter = value);
-    final prefs = await SharedPreferences.getInstance();
-    if (value == null) {
-      await prefs.remove(TrainingPackTemplatePrefs.streetFilter);
-    } else {
-      await prefs.setString(TrainingPackTemplatePrefs.streetFilter, value);
-    }
   }
 
   Future<void> _loadRecent() async {
@@ -550,38 +371,6 @@ class _TrainingPackTemplateListScreenState
     setState(() => _recentIds.clear());
     await prefs.remove(TrainingPackTemplatePrefs.recentPacks);
   }
-
-  bool _matchStack(int stack) {
-    final r = _stackFilter;
-    if (r == null) return true;
-    if (r.endsWith('+')) {
-      final min = int.tryParse(r.substring(0, r.length - 1)) ?? 0;
-      return stack >= min;
-    }
-    final parts = r.split('-');
-    if (parts.length == 2) {
-      final min = int.tryParse(parts[0]) ?? 0;
-      final max = int.tryParse(parts[1]) ?? 0;
-      return stack >= min && stack <= max;
-    }
-    return true;
-  }
-
-  String _streetLabel(String? street) {
-    switch (street) {
-      case 'preflop':
-        return 'Preflop Focus';
-      case 'flop':
-        return 'Flop Focus';
-      case 'turn':
-        return 'Turn Focus';
-      case 'river':
-        return 'River Focus';
-      default:
-        return 'Any Street';
-    }
-  }
-
   String _streetName(String street) {
     switch (street) {
       case 'preflop':
@@ -605,48 +394,7 @@ class _TrainingPackTemplateListScreenState
     return parts.join(' • ');
   }
 
-  String _filterSummary() {
-    final parts = <String>[];
-    if (_difficultyFilter != null) parts.add(_difficultyFilter!);
-    if (_posFilter != null) parts.add(_posFilter!.label);
-    if (_stackFilter != null) parts.add('${_stackFilter}bb');
-    String sortLabel() {
-      switch (_sort) {
-        case 'coverage':
-          return 'Coverage';
-        case 'created':
-          return 'Newest';
-        case 'spots':
-          return 'Most Spots';
-        case 'tag':
-          return 'Tag';
-        case 'last_trained':
-          return 'Last Trained';
-        default:
-          return 'Name';
-      }
-    }
-    parts.add('Sort: ${sortLabel()}');
-    return parts.join(' • ');
-  }
 
-  Future<void> _resetFilters() async {
-    setState(() {
-      _difficultyFilter = null;
-      _posFilter = null;
-      _stackFilter = null;
-      _streetFilter = null;
-      _selectedTag = null;
-      _tagFilters.clear();
-      _icmOnly = false;
-      _completedOnly = false;
-    });
-    final prefs = await SharedPreferences.getInstance();
-    await prefs.remove(TrainingPackTemplatePrefs.difficultyFilter);
-    await prefs.remove(TrainingPackTemplatePrefs.posFilter);
-    await prefs.remove(TrainingPackTemplatePrefs.stackFilter);
-    await prefs.remove(TrainingPackTemplatePrefs.streetFilter);
-  }
 
   List<String> _topTags(TrainingPackTemplate tpl) {
     final counts = <String, int>{};
@@ -730,143 +478,6 @@ class _TrainingPackTemplateListScreenState
         );
       }
     }
-  }
-
-  Widget _buildRecentCard(TrainingPackTemplate t) {
-    final total = t.spots.length;
-    final ratio = total == 0 ? 0.0 : (_progress[t.id]?.clamp(0, total) ?? 0) / total;
-    return Container(
-      width: 160,
-      margin: const EdgeInsets.symmetric(horizontal: 8),
-      padding: const EdgeInsets.all(8),
-      decoration: BoxDecoration(
-        color: Colors.grey[850],
-        borderRadius: BorderRadius.circular(8),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(t.name, maxLines: 1, overflow: TextOverflow.ellipsis),
-          const SizedBox(height: 8),
-          LinearProgressIndicator(value: ratio),
-          const Spacer(),
-          Align(
-            alignment: Alignment.bottomRight,
-            child: IconButton(
-              icon: const Icon(Icons.play_arrow),
-              onPressed: () => _chooseVariant(t),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildRecentSection(List<TrainingPackTemplate> list) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 16),
-          child: Row(
-            children: [
-              const Text('Recent Packs'),
-              const Spacer(),
-              TextButton(onPressed: _clearRecent, child: const Text('Clear')),
-            ],
-          ),
-        ),
-        SizedBox(
-          height: 120,
-          child: ListView(
-            scrollDirection: Axis.horizontal,
-            children: [for (final t in list) _buildRecentCard(t)],
-          ),
-        ),
-      ],
-    );
-  }
-
-  bool _isIcmTemplate(TrainingPackTemplate t) {
-    if (t.meta['icmType'] != null) return true;
-    if (t.spots.isEmpty) return false;
-    for (final s in t.spots) {
-      if (!s.tags.contains('icm')) return false;
-    }
-    return true;
-  }
-
-  List<TrainingPackTemplate> _visibleTemplates() {
-    final base = _showFavoritesOnly
-        ? [
-            for (final t in _templates)
-              if (_favorites.contains(t.id) ||
-                  FavoritePackService.instance.isFavorite(t.id))
-                t
-          ]
-        : _templates;
-    final byType = _selectedType == null
-        ? base
-        : [for (final t in base) if (t.gameType == _selectedType) t];
-    final filtered = _selectedTag == null
-        ? byType
-        : [for (final t in byType) if (t.tags.contains(_selectedTag)) t];
-    final icmFiltered = !_icmOnly
-        ? filtered
-        : [for (final t in filtered) if (_isIcmTemplate(t)) t];
-    final stackFiltered = _stackFilter == null
-        ? icmFiltered
-        : [for (final t in icmFiltered) if (_matchStack(t.heroBbStack)) t];
-    final posFiltered = _posFilter == null
-        ? stackFiltered
-        : [for (final t in stackFiltered) if (t.heroPos == _posFilter) t];
-    final diffFiltered = _difficultyFilter == null
-        ? posFiltered
-        : [
-            for (final t in posFiltered)
-              if (t.difficulty == _difficultyFilter ||
-                  t.tags.contains(_difficultyFilter!))
-                t
-          ];
-    final streetFiltered = _streetFilter == null
-        ? diffFiltered
-        : [for (final t in diffFiltered) if (t.targetStreet == _streetFilter) t];
-    final evalFiltered = !_showNeedsEvalOnly
-        ? streetFiltered
-        : [
-            for (final t in streetFiltered)
-              if (t.evCovered < t.totalWeight ||
-                  t.icmCovered < t.totalWeight)
-                t
-          ];
-    final completed = _completedOnly
-        ? [for (final t in evalFiltered) if (t.goalAchieved) t]
-        : evalFiltered;
-    final visible = _hideCompleted
-        ? [
-            for (final t in completed)
-              if ((t.spots.isEmpty
-                      ? 0.0
-                      : (_progress[t.id] ?? 0) / t.spots.length) < 1.0 ||
-                  !t.goalAchieved)
-                t
-          ]
-        : completed;
-    final inProgressFiltered = !_showInProgressOnly
-        ? visible
-        : [
-            for (final t in visible)
-              if ((_stats[t.id]?.lastIndex ?? 0) > 0 &&
-                  (_stats[t.id]?.accuracy ?? 1.0) < 1.0)
-                t
-          ];
-    return [
-      for (final t in inProgressFiltered)
-        if ((_query.isEmpty || t.name.toLowerCase().contains(_query)) &&
-            (_tagFilters.isEmpty ||
-                t.tags.any((tag) => _tagFilters.contains(tag))))
-          t
-    ];
   }
 
   void _scheduleAutoEval() {
@@ -2837,7 +2448,13 @@ class _TrainingPackTemplateListScreenState
                     subtitle: Text('$pct% • ${s.street} • $tags'),
                   );
                 }),
-                if (recent.isNotEmpty) _buildRecentSection(recent),
+                if (recent.isNotEmpty)
+                  RecentTrainingPackSection(
+                    templates: recent,
+                    progress: _progress,
+                    onClear: _clearRecent,
+                    onPlay: _chooseVariant,
+                  ),
                 if (history.isNotEmpty)
                   ExpansionTile(
                     title: const Text('Recent Generated Packs'),
