+++ lib/screens/v2/training_pack_template_prefs.dart
@@ -1,0 +1,21 @@
+class TrainingPackTemplatePrefs {
+  static const hideCompleted = 'tpl_hide_completed';
+  static const groupByStreet = 'tpl_group_by_street';
+  static const groupByType = 'tpl_group_by_type';
+  static const mixedHandGoalOnly = 'tpl_mixed_handgoal_only';
+  static const mixedCount = 'tpl_mixed_count';
+  static const mixedStreet = 'tpl_mixed_street';
+  static const mixedAuto = 'tpl_mixed_auto';
+  static const endlessDrill = 'tpl_endless_drill';
+  static const favorites = 'tpl_favorites';
+  static const showFavoritesOnly = 'tpl_show_fav_only';
+  static const sortOption = 'tpl_sort_option';
+  static const stackFilter = 'tpl_stack_filter';
+  static const posFilter = 'tpl_pos_filter';
+  static const difficultyFilter = 'tpl_diff_filter';
+  static const streetFilter = 'tpl_street_filter';
+  static const recentPacks = 'tpl_recent_packs';
+  static const inProgress = 'tpl_in_progress';
+  static const mixedLastRun = 'tpl_mixed_last_run';
+  static const continueLast = 'tpl_continue_last';
+}
